-tour reimport submenu in raw data and tourbook view -changed tourmarker distance from int to float -fixed min/max + y-slider bugs in tour chart for floating point values -changed data structure in SerieData
there are still endless bugs bug fixing the min/max bugs in the
tourchart was the most challenging task

+++ net.tourbook.device.garmin.fit/src/net/tourbook/device/garmin/fit/listeners/LapMesgListenerImpl.java
@@ -29,7 +29,7 @@ public class LapMesgListenerImpl extends AbstractMesgListener implements LapMesg
 
 		Float totalDistance = mesg.getTotalDistance();
 		if (totalDistance != null) {
-			int lapDistance = context.getLapDistance();
+			float lapDistance = context.getLapDistance();
 			lapDistance += DataConverters.convertDistance(totalDistance);
 			context.setLapDistance(lapDistance);
 			getTourMarker().setDistance(lapDistance);
